Correction d'un type  (passage de int à float pour la durée des déplacements)
Cela entrainait des itinéraires pareto-optimaux très proches particulièrement pour le mode vélo

+++ lib/core/graph_wrapper.h
@@ -66,7 +66,7 @@ void serialize(Archive &ar, Time &t, const unsigned int version)
 
 struct Duration
 {
-    int const_duration;
+    float const_duration;
     std::vector<Time> timetable;
     Duration();
     Duration(float const_duration);
